# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Generic SNAP/ASPI-compliant payment gateway API: B2B access\ntoken issuance, admin client onboarding, signature-generation\nutilities, and Virtual Account (transfer-VA) transaction and\nmerchant-dashboard endpoints.",
        "title": "SNAP Payment Integration Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/utilities/signature-auth": {
            "post": {
                "description": "Generates the SHA256withRSA signature required as X-SIGNATURE when calling POST /openapi/v1.0/access-token/b2b.",
                "tags": [
                    "Signature Utilities"
                ],
                "summary": "Generate the asymmetric access-token signature",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client identifier",
                        "name": "X-CLIENT-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RSA private key (PEM) used to sign the string-to-sign",
                        "name": "Private_Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureAuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureAuthResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureAuthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/utilities/signature-service": {
            "post": {
                "description": "Generates the HMAC-SHA512 signature required as X-SIGNATURE when calling SNAP-protected transaction endpoints (transfer-VA, merchant VA dashboard) with an issued access token.",
                "tags": [
                    "Signature Utilities"
                ],
                "summary": "Generate the symmetric service (transaction) signature",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client secret shared with the partner",
                        "name": "X-CLIENT-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HTTP method of the target request, e.g. POST",
                        "name": "HttpMethod",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative path of the target request, e.g. /openapi/v1.0/transfer-va/inquiry",
                        "name": "EndpointUrl",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer access token issued by POST /openapi/v1.0/access-token/b2b",
                        "name": "AccessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Raw request body of the target request, used verbatim in the string-to-sign",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureServiceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureServiceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureServiceResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SignatureServiceResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/v1.0/access-token/b2b": {
            "post": {
                "security": [
                    {
                        "SnapClientKey": []
                    },
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Issues a bearer access token for a registered client, after verifying the asymmetric X-SIGNATURE against the client's registered public key. This endpoint ISSUES the token — the caller has no prior bearer token; auth is via the X-CLIENT-KEY/X-TIMESTAMP/X-SIGNATURE headers only. Use POST /api/v1/utilities/signature-auth to compute X-SIGNATURE.",
                "tags": [
                    "Token"
                ],
                "summary": "Issue a SNAP B2B access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client identifier issued at onboarding",
                        "name": "X-CLIENT-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601 (e.g. 2026-07-24T10:00:00+07:00)",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asymmetric signature; compute via POST /api/v1/utilities/signature-auth",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware for all non-GET requests on this route. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Token request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload, missing required SNAP headers, or missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Signature verification failed",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SNAPTokenResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.SNAPTokenRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "grantType": {
                    "type": "string"
                }
            }
        },
        "domain.SNAPTokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "expiresIn": {
                    "type": "string"
                },
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
        "domain.SignatureAuthResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "domain.SignatureServiceResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Static admin API key configured via the ADMIN_API_KEY environment variable. Required for all /admin/* endpoints.",
            "type": "apiKey",
            "name": "X-Admin-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer access token issued by POST /openapi/v1.0/access-token/b2b, e.g. \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SnapClientKey": {
            "description": "Client identifier issued by the admin during onboarding. Required on the B2B access-token endpoint.",
            "type": "apiKey",
            "name": "X-CLIENT-KEY",
            "in": "header"
        },
        "SnapSignature": {
            "description": "Request signature. Compute via POST /api/v1/utilities/signature-auth (asymmetric, access-token endpoint) or POST /api/v1/utilities/signature-service (symmetric, transaction endpoints).",
            "type": "apiKey",
            "name": "X-SIGNATURE",
            "in": "header"
        },
        "SnapTimestamp": {
            "description": "Request timestamp in ISO 8601 format (e.g. 2026-07-24T10:00:00+07:00).",
            "type": "apiKey",
            "name": "X-TIMESTAMP",
            "in": "header"
        }
    }
}